Push local updates

+++ contracts/btc_nonced_output_claim_handler/BitcoinNoncedOutputClaimHandler.sol
@@ -27,7 +27,7 @@ contract BitcoinNoncedOutputClaimHandler {
 
         //Commitment
         bytes32 expectedTxoHash; //32-bytes
-        uint256 confirmations; //4-bytes
+        uint32 confirmations; //4-bytes
         address btcRelayContract; //20-bytes
 
         bytes32 commitmentHash;
@@ -45,9 +45,9 @@ contract BitcoinNoncedOutputClaimHandler {
 
         //Witness
         StoredBlockHeader memory blockheader = StoredBlockHeaderImpl.fromCalldata(witness, 56); //160-bytes
-        uint256 vout; //4-bytes
+        uint32 vout; //4-bytes
         bytes memory rawTransaction;
-        uint256 position; //4-bytes
+        uint32 position; //4-bytes
         bytes32[] calldata proof;
         assembly ("memory-safe") {
             let calldataPtr := add(witness.offset, 216)
@@ -72,11 +72,11 @@ contract BitcoinNoncedOutputClaimHandler {
         //Parse transaction
         BitcoinTx memory transaction = BitcoinTxImpl.fromMemory(rawTransaction);
         //Check output is valid
-        uint256 outputValue = transaction.getOutputValue(vout);
+        uint64 outputValue = transaction.getOutputValue(vout);
         bytes32 scriptHash = transaction.getOutputScriptHash(vout);
 
         //Get the tx locktime
-        uint256 locktimeSub500M = transaction.getLocktime() - 500_000_000;
+        uint32 locktimeSub500M = transaction.getLocktime() - 500_000_000;
         
         //Check the nSequence is correct
         uint256 firstNSequence = transaction.getInputNSequence(0);
