+++ contracts/btc_relay/utils/Difficulty.sol
@@ -8,15 +8,15 @@ library Difficulty {
     //Lowest possible mining difficulty - highest possible target
     uint256 internal constant UNROUNDED_MAX_TARGET = 0x00000000FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF;
     uint256 internal constant ROUNDED_MAX_TARGET = 0x00000000FFFF0000000000000000000000000000000000000000000000000000;
-    uint256 internal constant ROUNDED_MAX_TARGET_NBITS = 0xFFFF001D;
+    uint32 internal constant ROUNDED_MAX_TARGET_NBITS = 0xFFFF001D;
 
     //Bitcoin epoch timespan
     uint256 internal constant TARGET_TIMESPAN = 14 * 24 * 60 * 60; //2 weeks
     uint256 internal constant TARGET_TIMESPAN_DIV_4 = TARGET_TIMESPAN / 4;
     uint256 internal constant TARGET_TIMESPAN_MUL_4 = TARGET_TIMESPAN * 4;
 
-    function computeNewTarget(uint256 prevTimestamp, uint256 startTimestamp, uint256 prevTarget) pure internal returns (uint256 newTarget) {
-        uint256 timespan = prevTimestamp - startTimestamp;
+    function computeNewTarget(uint32 prevTimestamp, uint32 startTimestamp, uint256 prevTarget) pure internal returns (uint256 newTarget) {
+        uint256 timespan = uint256(prevTimestamp) - uint256(startTimestamp);
         
         //Difficulty increase/decrease multiples are clamped between 0.25 (-75%) and 4 (+300%)
         if(timespan < TARGET_TIMESPAN_DIV_4) timespan = TARGET_TIMESPAN_DIV_4;
@@ -26,8 +26,8 @@ library Difficulty {
         if(newTarget > UNROUNDED_MAX_TARGET) newTarget = UNROUNDED_MAX_TARGET;
     }
 
-    function computeNewTargetAlt(uint256 prevTimestamp, uint256 startTimestamp, uint256 prevReversedNbits, bool clampTarget) pure internal returns (uint256 newTarget, uint256 newReversedNbits) {
-        uint256 timespan = prevTimestamp - startTimestamp;
+    function computeNewTargetAlt(uint32 prevTimestamp, uint32 startTimestamp, uint32 prevReversedNbits, bool clampTarget) pure internal returns (uint256 newTarget, uint32 newReversedNbits) {
+        uint256 timespan = uint256(prevTimestamp) - uint256(startTimestamp);
         //Difficulty increase/decrease multiples are clamped between 0.25 (-75%) and 4 (+300%)
         if(timespan < TARGET_TIMESPAN_DIV_4) timespan = TARGET_TIMESPAN_DIV_4;
         if(timespan > TARGET_TIMESPAN_MUL_4) timespan = TARGET_TIMESPAN_MUL_4;
